LadspaEffectOptionsDialog defined in a header
