+++ src/CMakeLists.txt
@@ -538,6 +538,7 @@ list( APPEND SOURCES
          effects/ladspa/LadspaEditor.h
          effects/ladspa/LadspaEffect.cpp
          effects/ladspa/LadspaEffect.h
+	 effects/ladspa/LadspaEffectOptionsDialog.h
          effects/ladspa/LadspaEffectsModule.cpp
          effects/ladspa/LadspaEffectsModule.h
          effects/ladspa/LadspaInstance.cpp
